Model output instruction files
Here are an example of model output files used to read data into PEST.
The "modout.ins" is the instruction file and the "modout.out" is the
file to be read using the instruction file. Use the PEST utility
function "inschek.exe" to check if the instruction file is working the
way you intended. after you get the ins file together and check it, we
could do a PEST run